lab 16 - 17

+++ src/main/java/com/axac/ApplicationUser/Repositoris/UserRepository.java
@@ -4,6 +4,6 @@ package com.axac.ApplicationUser.Repositoris;
 import com.axac.ApplicationUser.Models.ApplicationUser;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<ApplicationUser,Integer> {
+public interface UserRepository extends CrudRepository<ApplicationUser,Long> {
     ApplicationUser findByUsername(String username);
 }
